Updating multi-run plots
